+++ apps/app/src/features/workflow/client/stores/workflow.ts
@@ -1,0 +1,29 @@
+
+import { SWRResponse } from 'swr';
+
+import { useStaticSWR } from '~/stores/use-static-swr';
+
+export type WorkflowModalStatus = {
+  pageId?: string,
+  isOpened: boolean,
+}
+
+type WorkflowModalUtils = {
+  open(pageId: string): void,
+  close(): void,
+}
+
+export const useWorkflowModal = (): SWRResponse<WorkflowModalStatus, Error> & WorkflowModalUtils => {
+
+  const initialStatus: WorkflowModalStatus = { isOpened: false };
+  const swrResponse = useStaticSWR<WorkflowModalStatus, Error>('workflowModal', undefined, { fallbackData: initialStatus });
+
+  return Object.assign(swrResponse, {
+    open: (pageId: string) => {
+      swrResponse.mutate({ isOpened: true, pageId });
+    },
+    close: () => {
+      swrResponse.mutate({ isOpened: false });
+    },
+  });
+};
