+++ apps/app/src/stores/modal.tsx
@@ -739,31 +739,3 @@ export const useLinkEditModal = (): SWRResponse<LinkEditModalStatus, Error> & Li
     },
   });
 };
-
-/*
-* workflowModal
-*/
-export type WorkflowModalStatus = {
-  pageId?: string,
-  isOpened: boolean,
-}
-
-type WorkflowModalUtils = {
-  open(pageId: string): void,
-  close(): void,
-}
-
-export const useWorkflowModal = (): SWRResponse<WorkflowModalStatus, Error> & WorkflowModalUtils => {
-
-  const initialStatus: WorkflowModalStatus = { isOpened: false };
-  const swrResponse = useStaticSWR<WorkflowModalStatus, Error>('workflowModal', undefined, { fallbackData: initialStatus });
-
-  return Object.assign(swrResponse, {
-    open: (pageId: string) => {
-      swrResponse.mutate({ isOpened: true, pageId });
-    },
-    close: () => {
-      swrResponse.mutate({ isOpened: false });
-    },
-  });
-};
